copy out search result
